so good

+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: hakader <hakader@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/20 11:54:57 by hakader           #+#    #+#             */
-/*   Updated: 2025/06/21 15:00:21 by hakader          ###   ########.fr       */
+/*   Updated: 2025/06/21 18:44:18 by hakader          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ typedef struct s_philo
 	int				meals_eaten;
 	int				left_fork;
 	int				right_fork;
-	int				last_meal_time;
+	long long		last_meal_time;
 	pthread_mutex_t	meal_mutex;
 	pthread_t		thread_id;
 	t_data			*data;
@@ -52,13 +52,13 @@ typedef struct s_data
 	pthread_mutex_t	stop_mutex;
 	pthread_mutex_t access_mutex;
 	int				stop;
-	int				start_time;
+	long long		start_time;
 }	t_data;
 
 // Utility functions
 int		ft_atoi(const char *str);
 void	put_error(char *msg);
-int		current_time(void);
+long long		current_time(void);
 void	print_action(t_philo *philo, char *msg);
 void *philo_routine(void *arg);
 void *monitor_routine(void *arg);
